+++ CHANGELOG.md
@@ -7,6 +7,34 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 
 ## Unreleased
 
+### Added
+
+* timeline markers - user can have more control over markers that are rendered on the timeline. See `TimelineMarkers` section of README for documentation - #327
+
+### Breaking
+
+* Removed support for React 15 and lower.  This is due to the fact that 16+ supports returning arrays from render, something that the TimelineMarker feature relies on.
+* removed `showCursorLine` prop in favor of using the `CursorMarker` component. See `TimelineMarkers` section of README for documentation.
+
+```diff
+import Timeline,
++ {TimelineMarkers, CursorMarker}
+from 'react-calendar-timeline'
+
+<Timeline
+- showCursorLine
+
+- />
++ >
+
++ <TimelineMarkers>
++  <CursorMarker />
++ </TimelineMarkers>
+
++ </Timeline>
+```
+
+
 
 ### 0.17.2
 
@@ -21,6 +49,7 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 ### Added
 
 * pass canvasTimeStart/End via timelineContext to the itemRenderer prop
+
 
 ### 0.17.0